move HashMessage out of the backend struct
