+++ api/backend_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/status-im/status-go/node"
 	"github.com/status-im/status-go/params"
 	"github.com/status-im/status-go/rpc"
-	"github.com/status-im/status-go/services/personal"
 	"github.com/status-im/status-go/t/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -382,36 +381,4 @@ func TestStartStopMultipleTimes(t *testing.T) {
 	require.NoError(t, backend.StopNode())
 	require.NoError(t, backend.StartNode(config))
 	require.NoError(t, backend.StopNode())
-}
-
-func TestHashMessage(t *testing.T) {
-	backend := NewStatusBackend()
-	config, err := utils.MakeTestNodeConfig(params.StatusChainNetworkID)
-	require.NoError(t, err)
-	err = backend.StartNode(config)
-	require.NoError(t, err)
-	defer func() {
-		require.NoError(t, backend.StopNode())
-	}()
-
-	key, err := crypto.GenerateKey()
-	require.NoError(t, err)
-	addr := crypto.PubkeyToAddress(key.PublicKey)
-
-	originalMessage := []byte{0x01, 0x02, 0x03}
-	hash := backend.HashMessage(originalMessage)
-
-	// simulate signature from external signer like a keycard
-	sig, err := crypto.Sign(hash, key)
-	require.NoError(t, err)
-	sig[64] += 27
-
-	// check that the message was wrapped correctly before hashing it
-	recParams := personal.RecoverParams{
-		Message:   fmt.Sprintf("0x%x", originalMessage),
-		Signature: fmt.Sprintf("0x%x", sig),
-	}
-	recoveredAddr, err := backend.Recover(recParams)
-	require.NoError(t, err)
-	assert.Equal(t, addr, recoveredAddr)
 }
