+++ lib/library.go
@@ -458,14 +458,14 @@ func HashTransaction(txArgsJSON *C.char) *C.char {
 // The hash is calulcated as
 //   keccak256("\x19Ethereum Signed Message:\n"${message length}${message}).
 // This gives context to the signed message and prevents signing of transactions.
-//export HashTransaction
+//export HashMessage
 func HashMessage(messageString *C.char) *C.char {
 	message, err := hex.DecodeString(C.GoString(messageString))
 	if err != nil {
 		return C.CString(prepareJSONResponseWithCode(nil, err, codeFailedParseParams))
 	}
 
-	hash := statusBackend.HashMessage(message)
+	hash := api.HashMessage(message)
 	return C.CString(prepareJSONResponseWithCode(fmt.Sprintf("0x%x", hash), err, codeUnknown))
 }
